requirements.txt
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0  # For making HTTP requests to AccuWeather API 